wip: parse mml

+++ lib_player/src/mml_event.rs
@@ -1,7 +1,7 @@
 use crate::note_event::NoteEvent;
 
 pub enum MmlEvent {
-    SetTempo(u16),
+    SetTempo(usize),
     SetVelocity(u8),
     SetOctave(u8),
     IncreOctave,
